+++ default.nix
@@ -17,6 +17,7 @@ ghc."8.10" = nixos-22-05.haskell.packages.ghc8107.override (old: {
     overrides = combineOverrides old [
         sourceOverrides
         (new: old: {
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.0.1.nix {};
         })
     ];
 });
@@ -25,9 +26,9 @@ ghc."9.0" = nixos-22-11.haskell.packages.ghc90.override (old: {
     overrides = combineOverrides old [
         sourceOverrides
         (new: old: {
-            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
-            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.nix {};
+            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.nix {};
         })
     ];
 });
@@ -36,9 +37,9 @@ ghc."9.2" = nixos-22-11.haskell.packages.ghc92.override (old: {
     overrides = combineOverrides old [
         sourceOverrides
         (new: old: {
-            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
-            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.1.0.0.nix {};
+            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.nix {};
+            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.0.nix {};
         })
     ];
 });
@@ -47,9 +48,9 @@ ghc."9.4" = nixos-22-11.haskell.packages.ghc94.override (old: {
     overrides = combineOverrides old [
         sourceOverrides
         (new: old: {
-            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.0.nix {};
-            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.0.nix {};
-            ascii-superset = new.callPackage ./nix/ascii-superset-1.2.0.0.nix {};
+            ascii-case = new.callPackage ./nix/ascii-case-1.0.1.nix {};
+            ascii-caseless = new.callPackage ./nix/ascii-caseless-0.0.0.nix {};
+            ascii-superset = new.callPackage ./nix/ascii-superset-1.3.0.nix {};
         })
     ];
 });